fix: Accounting for multiple results from lsof
When querying lsof from the python process this can provide multiple results
with the same pid. The second is from the lsof query itself (this was observed
with an earlier change for querying our own process). Accounting for this by
checking all pid results from the query to see if they're in agreement.