fix: Proc resolution without a pid gave stacktrace
When the connection resolver hadn't been configured with a pid and we queried
connections by proc resolution there was an uncaught exception. This didn't
crash arm - it just resulted in a stacktrace for a split second at startup.
This was most easily reproed by running arm on Linux in blind mode, with a
sleep command between panel creation and redrawing the screen to keep the stack
trace visible for a bit.