Delaying connection resolution until after init
This is a huge performance fix, dropping the startup time from 0.84 seconds to
0.14 (83% improvement). Arm had been fetching connection contents on init,
blocking the first redraw. The fix is to both move connection init into its
runtime thread *and* delay its execution until after the initial redraw. This
hides the call latency in the period that we're usually waiting for user input
anyway.

There's still perceived latency if the user immediately switches to the
connections page, but this is both unavoidable and unusual.