Moving help popup into a new popup toolkit
Breaking the help popup out of the controller, moving it to a nice, modular
'show popup' function. This is a drop-in replacement (besides some reordered
help options) and also allowed us to move the help content into their
respective panels.

This is far cleaner and avoids accessing private panel parameters. Next the
rest of the popups will be moved to this new toolkit, which will greatly clean
up the controller.