Dropping unused config switcher function
Once upon a time the config panel showed both tor and arm configuration
options. However, arm doesn't yet have a counterpart for SETCONF so I dropped
the arm config from the interface.

I'll probably re-introduce this at some point, but for now this is just dead
code and the back-end needs some more work before it's useful to users. Since
the showMenu function is being refactored this dead code would, as is, be
broken so dropping it.