fix: Never show conn details if panel's empty
The connection panel could get stuck in the 'show connection details' mode if
the panel's contents became empty while we were showing something. This
wouldn't crash (like with the issue addressed in the previous fix), but the
panel title would be wrong. This reverts the 'show details' flag if we don't
have any contents.