fix: dropping T-connector from config save dialog
Trying to get the config save dialog to display the right pipe connectors to
dynamically align with the below content is a pita so gonna stop trying.
There's three use cases...
- dialog is larger than the detail section, no scroll bars -> display box
- dialog is larger and there are scroll bars -> show T-connector on left
- dialog is smaller than the detail section -> show connectors on both sides

Trying to get enough context to the dialog really isn't worth the hassle, so
just opting to show a simple box in all situations.