Replaced namedtuples with MenuItem instances, fixed left arrow key on top level menu.
The MenuItem class is added replacing LeafEntry and ParentEntry
namedtuples. Replaced "entry" with "item" and defined the constant
TOPLEVEL to improve readability.

The scrolling calculations in _moveTopLevelLeft contained a bug, menus
with large number of top items did not scroll to the end. Fixed by
adding a check to see if we're "spilling" to left of first entry.