Mapped menu items to respective actions.
There are three kinds of menu items:
 * The ones which are directly mapped to a panel action via
   Menu._callbackPressKey()
 * "State-ful" menu items which are enabled/disabled depending on their
   panel's states. These are taken care of via the
   Menu._callbackGetEnabled() and Menu._callbackSet() methods. For
   example, the "Duplicates" submenu in "Logs" changes its state
   depending on log.showDuplicates.
 * View menu items, which change controller state. Supported via
   Menu._callbackView()

Launching a panel action switches view to that panel, e.g., if we're
on configuration panel and graph bounds are changed the controller
is traversed through pages until one with graph panel is found.
Otherwise, an error is logged.

File->Exit and Configuration->Reset Tor items are unimplemented as of
yet. Possible approaches include factoring out a handleKey() in
controller or duplicating the same code in callbacks.