Next top-level item is dropped down in case right/left arrow keys are pressed on leaf submenus.
In order to let the controller redraw before next top-level item is
dropped down, "menu keys" are calculated and returned back to the
controller. The controller overrides its key so that menu is redrawn,
redraws all other panels and sends back the "menu keys" to the new menu
instance. This maintains the illusion of top-level drop-down moving
right/left while screen is cleared and redrawn once in between.