Adding tarball dependencies for autofetching
The autofetched library dependencies weren't having their signatures checked,
making this a very stupid vulnerability when on an untrusted network (that
said, in practice you'd need to check arm's integrity for this fix to be
meaningful which isn't possible when it's coming from git). However, this is
probably better than nothing. Suggested by Sebastian and rransom