Improved interaction for graph resizing
Rather than using the 'n' and 'm' keys to resize the graph (with was both bad
and being overwritten by other functions) we now use the 'r' key to enter a
resizing mode where they can use up/down to resize the graph. This is more
intuitive and avoids occupying extra hotkeys.