fix: loop from closing conn based on torctl event
Attempting to close our control connection when TorCtl logged a message
saying that it was done caused us to attempt to join our own thread within
torctl (which causes an exception).

I'm not really recalling why I ever did this, but fixes for concurrency issues
in TorCtl are triggering this so removing the check.