fix: race condition for multiple status changes
When an event caused multiple changes in Tor's status we'd run into a race
condition and often pick the wrong final state. For instance, if a sighup
caused Tor to crash (for instance via a bad torrc) then we'd trigger both a
reset and closed event.

This enqueues events to keep the order in which they occured, expands locks
to include them, and introduces a small delay to coles events if they occure
in quick succession like this.