fix: Lowering get_network_status call memory usage
When generating the IP -> fingerprint mappings I call the TorCtl
get_network_status function which generates a TorCtl.NetworkStatus instance for
every relay that exists. Since all of the memory is allocated at once this
increases the preceived memory footprint of arm by 3.5 MB. This is because even
after the memory's freed the python allocator doesn't actually give the memory
back:
http://effbot.org/pyfaq/why-doesnt-python-release-the-memory-when-i-delete-a-large-object.htm

This uses (and relies) on a change in TorCtl to use an iterator instead to
avoid this mass allocation:
https://gitweb.torproject.org/atagar/pytorctl.git/commit/45d7753b7c947b141210017d1b987673c51d97c5

In practice this lowers arm's baseline memory usage by 2.5 MB (12%).