Dropping remaining addfstr usage
Long ago, in the dawn of arm, I added a helper function "addfstr" that allowed
for adding strings with formatting tags similar to html. All usage of it with
arbitrary input was dropped due to the possability of tag injection screwing
up the display, and this change drops its last remaining usage.

This hurts the code's readability a bit and at I'd like to figure out a nicer,
general purpose pattern for multi-format line rendering but tags aren't the
solution. Back to the drawing board...

This leaves the panel utility for now.