Adding a "New Identity" option
Providing an option for sending newnym signals to tor. This is both provided
via a menu option and a 'n' hotkey. There's also an indicator in the header
panel if we're both on a wide screen and running as a non-relay. This was
requested on 'https://trac.torproject.org/projects/tor/ticket/3394'.