Optional detached startup
This is adding a config option that allows us to run arm when there's no Tor
instance to connect to. This is a prerequisite for both a relay setup wizard
and arbitrary reattachability.

Arm was written with the expectation that there was a controller instance
while starting up. This commit fixes most of the obvious crashing issues, but
this is still gonna take some work to address instability for detached
startups.