Wizard config for setting up an exit
This expands the wizard to provide a config panel for setting up an exit relay.
This has the same options as an internal relay plus several for
DirPortFrontPage and ExitPolicy.

By default this will use the reduced exit policy [1], with options like Vidalia
for selecting groups of services to allow. However, unlike Vidalia it won't
allow the operator to pick plaintext ports without also having their encryped
counterparts.

[1] https://trac.torproject.org/projects/tor/wiki/doc/ReducedExitPolicy