Torrc generation from the wizard input
This is most of the functionality needed to generate a torrc from the relay
setup wizard's input (minus a few special-case fields like the exit policy
and burst rate).

This is done via a template that's filled in by the wizard results. The
templating language just contains the simple set of control structures (if,
if not, or) needed for this.