Handling when tor needs root to start
When connecting to privileged ports the tor process needs root permissions to
start. If the torrc needs root then I shouldn't offer to start tor with arm,
and the wizard makes a startup shell script rather than making a doomed
attempt to start tor.

Currently the startup script is just a stub - I'll need to wait until I have a
connection to fill in all the shell scripting voodoo.