fix: repopulating log panel when on tor init
When connecting to a new tor instance the log panel is now clearing itself and
repopulating with the event backlog. This keeps all arm events, and mixes them
with the events of the tor log file. This is helpful since it means that we'll
display the events that occurred during initialization.