Replace "interval updates" by consistent list->model mapping.
The CLI config panel stored entries in
ConfigPanel.confImportantContents which were regularly polled and
exported to GUI treestore.

This created problems with selections as the liststore was essentially
emptied at each timeout.

The new approach is to override ConfigPanel.confImportantContents
with a wrapper class which maintains consistency between the Python
list and GUI liststore at each append/setitem.