Python black-magic to connect lists with liststores.
The ConfigPanel.confImportantContents is patched via Python property()
in such a way that assigning lists to it automatically updates
the liststore. The setter/getter is really ugly and hard to read, and
serves only the purpose that the CLI portion does not need to be
changed at all for this to work.

Once library/CLI/GUI is decoupled this obscure piece of dark magic
shall be gone for good.