Replace list->treestore caching from conn panel with wrappers.
The old method of refreshing treestore entries every three seconds via
the CLI counterpart is replaced by the list wrapper (developed for
config panel, part of gtkTools).

The list wrapper patches the conn panel so that each time the CLI
panel updates entryLines, the relevant treestore structures are updated
automatically. This way, the treestores are only updated when arm
backend is updated.