fix: avoiding setresuid/gid if unavailable
The os.setresuid and os.setresgid functions are only available in Python 2.7
and later. Arm aims for 2.5 compatability so using os.setreuid/gid if running
a prior version. This, unfortunately, means that the saved uid is not reduced
which might be a vulnerability - hopefully Jake will know of an alternative if
this is a concern.