Periodically refreshing display
On some platforms the terminal gets into a screwed up state, displaying
scrambled content (I haven't reproduced this on my system, but seen it on an
Ubuntu netbook).

Redrawing the display, reguardless of if the content has changed every five
seconds so this glitches will correct themselves. Arm had previously done this
but I took it out when the display code had stabilized (to cut down on load).
However, this doesn't have a noticeable cpu cost so might as well put it back.

This also simplifies some of the redrawing logic, centralizing it in the
controller.