fix: quitting could cause unclean curses shutdown
If the user quit arm while we're processing tor events then there was a rather
common timing issue where curses would terminate, then a daemon thread
triggered a curses redraw. The result was that the terminal would be screwed
up until the user ran 'reset'.

This was because the daemon shutdown was outside of the curses wrapper
function. Changing the ordering so we shut down daemons before curses, and
setting a flag during keyboard interrupts to suppress curses redraws.