Wizard components for a startup option
This adds platform detection and a check for the system torrc option for the
wizard startup option. However, this still isn't enabled nor is the option
implemented (so this has no impact on users). I'm suspecting that system
startup is a bigger can of worms than I'd like to open right now, even if
restricted to Debian and Ubuntu platforms so gonna leave the rest unimplemented
for now...