fix: failing fingerprint lookups when DirPort
Inbound and outbound connection types were assuming that the ports they were
connected to was a relay's ORPort and using this to narrow results in
fingerprint lookups. However, we could be attached to a relay's DirPort in
which case the lookup failed (but when pressing enter there was only a single
possible relay result).

Instead only using the port to narrow the results when there's multiple
matches.