Skeleton terminal contorl interpretor
This provides a basic skeleton for a terminal-only interpretor of the control
port. This provides terminal coloring, formatting, and history scollback but it
not yet tied into the control port. The plan is that both the torInterpretor
and interpretorPanel will use the same backend for irc style interpretor
commands so people can use the interpretor in both the curses interface or
plain terminal.