Replacing displayed consensus policy with util
Using the ExitPolicy class' summary rather than the string from the consensus.
Unless I'm misunderstanding how the consensus policy is derived, these should
be equivilant. However, since this doesn't rely on the policy being in the
consensus this will be rendered the same for both new and old versions of tor.