Adding back the color prompt, with padded prompt
This is a hack to maintain a color prompt while sidestepping the
'http://bugs.python.org/issue12972' bug in *most* use cases. This pads the
prompt so the original issue doesn't manifest unless the user's input exceeds
around sixty characters or so. However, it introduces a new bug where tiny
terminals (smaller than the padded width) get screwed up.

This obviously isn't a great situation and it could be argued that a colored
prompt isn't worth these issues. However,  the problem use cases strike me as
being fairly rare (and I much prefer the green and bolded '>>>') so going with
this for now. Patches welcome if someone has a better solution.