Source flags for color prompt and hostname lookups
Adding source flags for portions of the interpretor that include substantial
tradeoffs. For color prompts there's bugs in small displays/huge input, and
hostname lookups add a second or two to the runtime of /info queries (otherwise
the lookup is instantanious).

Defaulting the hostname lookups to be disabled for now. There's plans for
having cached results with...
https://trac.torproject.org/projects/tor/wiki/doc/arm#CircuitDetails

though I haven't a clue if it'll ever be implemented.