fix: concurrency issue could crash from CIRC event
If we got a CIRC event while populating the cache of relays that we're attached
to we could reset _fingerprintsAttachedCache to None while we were in the
middle of populating it. Manipulation of this cache is supposed to be governed
by the connection lock, so fixing that for the CIRC event's reset.