Quiting wizard when the user presses 'q'
Users could cancel the wizard by prssing esc, but it's also intuitive to expect
'q' to do the same (since that's the keybinding to quit arm entirely).
Respecting this option too. This was requested by monochromec on...
https://trac.torproject.org/projects/tor/ticket/3995

This also fixes a minor bug where the config options dialog wouldn't respect a
cancel signal (instead looping on that dialog).