fix: dropping os check for startup wizard option
The startup wizard option is currently unimplemented and disabled in the
setting.cfg. However, there was an os compatability check left over from when I
was gonna implement it. That and its comment were confusing since the option's
not implemented so commenting it out with a more up to date expanation.