Toning down arm's notice when resolvers fail
When arm failed to query tor's connections it would give a short warning saying
so. This often spooked users so lowering this to be a NOTICE and giving more
information on both what happened and how to fix it (99% of the time this is a
simple permissions issue). Change requested by Sebastian