Improving tor detection for BSD platforms
Linux and BSD platforms (mac, openbsd, and freebsd) have distinct argument sets
for their ps resolvers. Taking into account for this for tor detection,
guessing the family based on uname results then trying both if we guess wrong.

Adding the '-a' argument to get results for all users thanks to Jordi Espasa
Clofent.

Tested this ps command on Mac and OpenBSD, and just guessing that FreeBSD will
behave the same.