Option for disabling ACS support
Some terminals have issues with alternate character support which is widely
used by arm for borders. Unfortunatley this makes the interface look pretty
crappy...
http://www.atagar.com/arm/images/acs_display_failure.png

Explaining the issue in the README and adding an option to manually remedy the
most obvious broken bits. ACS is used throughout arm so it's gonna take a much
more invasive change to allow the user to completely replace ACS usage (for
instance, we can't use the curses box() function...).

Thanks to Robert for explaining the issue!