Monkey patch for ACS disabling
Making manual ACS disabling transparent to the rest of the interface by monkey
patching the ord values of ACS entries to '+', '-', or '|'. This has the
advantage that the rest of the arm interface doesn't need to be aware of this
complication, however it doesn't address the builtin box() function usage,
which still makes most dialogs render incorrectly.