Ignoring default deb Log config option
Tor debs have their default logging configuration changed via a patch, as
per...
https://trac.torproject.org/projects/tor/ticket/4602

Dropping this option when querying for the custom options the user has set to
avoid giving a torrc validation error.

This introduces a minor bug because if the user has this exact entry with a
non-deb install, or the normal default log entry with a deb install, then arm
will report that the entry matches the default. However, that would be a rare
situation and this whole thing is a workaround for a tor bug so I'm fine with
leaving it alone for now. Fixing that would need some method for checking if
the instance we're connecting to was deb installed or not (I'm not sure how
we'd do that).

Issue was caught on a Tor Cloud image by koolfy