Prevent crash on startup when reading Feb 29th in logs
During prepopulation of log entries, arm tries to parse the date that
Tor gives it. Unfortunately Python is a little (read: very) silly and
refuses to accept that February 29th can be a valid day when information
about the year is missing. As a solution, we just pretend it is the year
2012 forever when parsing dates from logs (Tor doesn't add the year to
its date information).

Fixes bug 5265, first reported by an anonymous user from IRC.