Adding arm icon svg and .desktop file
Adding the same sort of icon svg that I used on the website. This is based on
'utilities-system-monitor.svg' from the Oxygen icon set, which is used by KDE
and under GPLv2 and CC v3 (A, SA).

Also adding a '.desktop' metadata file for the arm gui provided by Juan.