Better validation for path component of circuit-staus output
The 'GETINFO circuit-status' output has an *optional* third arguement that is a
path. This is a pita because it's essentially an optional positional arguement
that can have '=', so we need to differentiate it from the key=value entities
that follow.

This resulted in confusing 'Unable to determine fingerprint' warnings, and
seemed to cause a screwed up circuit listing. Ticket tracking this is...
https://trac.torproject.org/projects/tor/ticket/5267