Lowering unrecognized event type logging to INFO
Unrecognized event types are neither important nor something that a user can be
expected to fix. Actually, it's an issue for the controller library (torctl or
stem). Dropping these below the default runlevel.

Arm is currently complaining about SIGNAL and CONF_CHANGED not being
recognized. If torctl doesn't add support for these then this won't be
addressed until our move to stem.