Using stem's get_info() method
Stem's get_info() method is very similar to our own, the most notable
difference being that it doesn't have a 'suppressExc' argument. Instead if it
gets a default then exceptions are suppressed, and if it doesn't then they're
raised. This is far more intuitive usage than ours, but means that our callers
need to add the 'None' value we were defaulting to before.