Dropping unused GETINFO caching
Stem has its own GETINFO caching so we can now drop our own. One small
difference is that stem does not cache 'GETINFO address', so we're loosing
that. That said, we were listening to NOTICE level events to figure out when it
was invalidated which was more than a little wrong.