Replacing TorCtl descriptor calls
Oops, missing a use of our TorCtl controller connection. This was a kinda
screwy call - we were checking the 'down' attribute of a TorCtl Router class.
This attribute seems to be True if...

* the router lacks the Running flag
* is hibernating
* is Valid but lacks an observed bandwidth

I'm not quite sure why I ever used this attribute. Just checking for the
Running flag and calling it good. We can check for other criteria if it makes
sense for the use case later.