Dropping the interpretor panel
Since introducing the interpretor panel to arm the program's been plagued
with stability issues (https://trac.torproject.org/6439). I suspect that the
root cause is related to the readline module somehow being imported, though
I tried to keep that import away from the cli use cases.

The interpretor panel was a sweet idea and will live on as a separate
application. However, as a part of arm it has been surprisingly unused (I've
*NEVER* heard of someone making use of it). Oh well...