Using stem's conf util
Ok, let's get this out of the way first. This is a major overhaul. Arm's conf
util had its insidious clutch over just about every file in our codebase.
Stem's conf util introduces a new pattern for managing configurations that
turns out to be vastly cleaner.

I'll be amazed if this doesn't introduce some sort of regression but so far so
good. In the long run this'll be a big step toward having a simpler, more
manageable codebase.