Using stem's log util
We were using our own log utility rather than python's builtin logging
framework. Piggybacking on stem's log utility so we can drop ours (not strictly
the right thing to do since it makes our event's look like stem's, but it's one
and the same to users).

This drops logging configurability, which I've been tempted to do for years.
Users very rarely make an armrc, and when they do I've never heard of someone
tweaking the runlevels at which events are logged. It was an interesting idea,
but cluttered the code for no reason.

This also fixes our ability to log stem/arm events, which was broken by the
mass refactoring to move to TorCtl. This has revealed an oversight in our
earlier conf change.