Using stem's system util
Stem has some (not all) of our system utilities. The most interesting one is
the call() function. Stem provides a subprocess based implementation that's
pretty simple, while ours was os.popen() based and a tangled mess of hundreds
of lines.

The reason for that mess was that I wanted to support caching and piped
arguments (like "ps -aux | grep tor"). However, we only used the caching at
one spot in the *entire* codebase, and the piping on reflection was a bad idea.
This is code - we can filter our own results without relying on grep and egrep.

This required a few changes to the connection util (the heaviest user of grep)
and has a couple points of regression...

* The bandwidthStats panel no longer gets cached ps results - we need to figure
  out a cleaner method for deduplicating those system calls.

* We used the runtimes of our system calls to partly calculate arm's cpu usage.
  This was a bit of a hack but I'm not sure of a better way of including those
  invocations in our usage. This will require some more thought.