Dropping complicated disambiguation step from _getRelayFingerprint()
The _getRelayFingerprint() helper starts with very sensible approaches to
resolve addresses/ports to fingerprints. However, after it has checked if it's
us and if the consensus holds the answer it does some mind bogglingly weird
steps involving orconn-status and circuit-status.

These final disambiguation steps are horribly slow (it performs at one point a
GETINFO for every relay in the consensus) and, from my read, pointless. I doubt
this logic has ever been encountered and, if it has, it probably hasn't helped
the resolution. My suspicion is that this is leftover code from arms very first
version.