Dropping functionality to sighup via pkill
Once I had a phobia of deleting any code that could possibly prove to be useful
later. This is a great example: the initial version of arm could reset tor by
issuing a pkill to sighup. I later learned that this could also be done far
more simpily by issuing a 'SIGNAL RELOAD'.

So what did I do? I changed arm to use SIGNAL of course, but I also left the
pkill code around just in case I needed it in the future. The only advantage of
pkill is that we can sighup tor instances that lack a control port. This might
make its way into stem's process module someday if we ever have a need, but
arm's implementation is old and unused so it's about time it got deleted.