Dropping unused setOptions() method
Arm's Controller was meant to be a complete wrapper around TorCtl, and as such
it provided some funtionality that arm itself doesn't use. One instance of this
is setOptions() which allowed the Controller to set multiple configuration
options at once. There's no method for doing this in the curses interface so
dropping the method.