Dropping getControllerEvents() method
The getControllerEvents() method was broken because arm's Controller no longer
tracks event listening. On reflection though it's also unneeded since it's sole
caller is checking for BW events. The graph panel requres those so they'll
always be present unless the user opted to drop the graph (possible, but not
done in practice).

This also cleans up some unused attributes and imports.