Renaming 'src' to 'arm'
Changing our 'src' directory to 'arm'. The old name is largely a holdover from
when I was a java developer, in the python world it seems more common to name
the source directory after your project.

One gotcha is that this means we can't have an 'arm' executable at the base
directory any longer since this would conflict with the directory. Pity.
Changing that to 'run_arm' for now, I might change it again later.