Shifting to a proper package layout
Making a couple important changes so we have a more proper layout for an arm
package...

* eliminating the 'cli' directory (legacy from when we had a gui)
* proper absolute imports (we were doing all imports relative of our base)

This no doubt breaks our installation process a thousand different ways, and
that will need an overhaul later. That said, ever since arm's start we've had
a very un-pythonic hacked up layout due to the above issues (and our 'src'
directory). This gets us a lot closer to having a proper python package!