Alternate function for parsing arguments
Adding a _get_args() function for parsing commandline arguments and providing
back a named tuple. This is similar to what I'm doing for stem's run_tests.py
argument parsing, and allows for much cleaner code than the present config
mess.

This also includes unit tests for the new function. My plan is to add tests as
I incrementally rewrite arm, but being a curses app I'm not sure if we'll be
able to achieve too much coverage. Still, any bit helps.

The _get_args() function is not yet used. It's gonna take a bit of work to
untangle the present config starter parameters.