Replacing _getController() with revised counterpart
Huh. Well, that was odd. We acquired control sockets in the main loop, but then
used a helper to both get and authenticate control ports. The sockets never
went through non-basic authentication. Meh, whatever.

Made a _get_controller() helper that... well, just gets a Controller. We then
authenticate reguardless of the connection method.