Removing sysTools' getFileErrorMsg() function
The sole purpose of getFileErrorMsg() was to strip off the errno component of
IOErrors from the output we print or log. On reflection however, that's exactly
what the IOError's strerror attribute is so might as well use this instead.
Only funcitonal difference of this change is that we no longer call lower() on
the messages.