Removing unused REQ_EVENTS dict
Our old TorCtl-based codebase needed to play a funky dance to get the event
types we needed, but also allow users to customize what they wanted for the log
panel. Stem has a much better API in this reguard so we can just add separate
listeners for both and it hanles the SETEVENT calls for us.

The REQ_EVENTS is legacy from those old, dark times. We no longer use it so it
can be safely removed.