Moving connection resolution init into initController()
Shifting where we initialize our connection resolver and provide a notice about
DisableDebuggerAttachment. I'm not a fan of this change, but I'm trying to
eliminate the start_arm() function. This is kicking the can down the road for
this chunk of code, but hopefully by then we'll be able to handle
DisableDebuggerAttachment.