Dropping hostnames module
In an earlier version we supported hostname resolution, but this leaked tor
connecion information to our DNS resolver (very, very bad). Since then we've
had the util for performing lookups but it's unused by the codebase.

I'd love to add a capability like this back in when we can do so safely, but
there's little point in keeping an unused module.