Removing blind mode
Blind mode is mostly a legacy of arguments around 'is connecion data
eavesdroppign or not?'. This argument was largely resolved my making arm
smarter about what constitutes a client or exit connection, and scrubbing
those.

Dropping the '--blind' argument. There could be good reasons to support
removal of connection lookups (for instance, to reduce system load) but it
shouldn't be this kind of first class argument.