Fetching string resources via a msg() helper
Over time I've been moving more and more strings to our config. This moves the
remainder of the starter strings over, places them in namespaces, and added a
helper to make our code for fetching string resources *much* nicer.

This might also help with internationalization later on if we go that route
(since strings are now being segregated to their own config).