Finish moving subtasks out of main()
Completing the exodus of subtasks from the main method to shorten it. This is
nice since main() now has short, human readable names for the subtasks, with
the implementations defined below. Tasks that remain either print or call exit
(it's bad form for functions to do that on their own).

This is about as nice as I can presently get the starter so it's about time I
shift my focus to other modules.