Dropping the sysTools module
The last function of arm's sysTools module was getSysCpuUsage() which was
broken by our move to stem.

When we revise the header panel we'll need to address this by adding a wrapper
around stem's call() function to get its runtimes. For how just leaving this as
a todo note, though.