Replacing uiTools isUnicodeAvailable()
Cleaning up our uiTools function for determining if curses supports unicode or
not. We don't need the caching of isUnicodeAvailable() nor did it help with
readability so dropping that. We still need the function for determining if
curses has wide character support though so just tweaking that to be PEP8
compliant.