Dropping tor_tools' get_my_file_descriptor_limit()
In Tor version 0.2.3.x-final we added the ability to query available tor's file
descriptor limit. This method was a holdover from before this was commonly
available, taking a guess at the limit based on either 'ulimit -Hn' or tor's
user.

This tor capability has been out there long enough that we can drop this
fallback functionality.