Dropping the notion of static header parameters
The header panel split the attributes it used into two camps: static and
volatile. The reason for this is that we wanted to avoid repeatedly querying
things that rarely change. However, the stem controller caches static values
for us anyway so we can drop this.

The only additional burden this is adding is for the process' start time. This
is now fetched on each call - haven't yet decided how we want to handle those.