Moving header panel sampling to its own class
Our header panel is a daemon thread that periodically pulls stats, rending the
last result whenever redrawn. We've changed what we use for sampling a few
times. First it was a dictionary, then a named tuple. Now I'm making the
sampling its own class. This greatly helps split the model and view
functionality.