Rewriting header panel's ports section
Overhauling the second header panel section.

For most (but not all) attributes we displayed either a full entry or nothing
at all. That is to say, if we had room for 'Dir Port: 80' then we show it, but
otherwise omit it entirely. This is generally preferable for really, really
tiny screens.

I'm dropping that. It made the code quite a bit more clunky and generally only
arises with consoles skinnier than fourty columns or so. That's... rare to say
the least, and it's perfectly fine for us to render with normal cropping there.