Don't attempt to show disconnection notice in the flags section
We showed identical disconnection notices in either the ports or flags section
depending on if we were a relay or not. This was based on us remembering our
ORPort from before the disconnection but we presently don't do that, so the
later was dead code.

Also moving the disconnection notice to its own draw function.