Rewriting header panel's exit policy section
This is gonna take some more work. The last release of arm consulted our
GETCONF ExitPolicy which excluded both the ExitPolicyRejectPrivate prefix and
default suffix.

We definitely want to condense the default suffix. As for the prefix, that's
probably just extra noise. We could just snag the GETCONF like arm previously
did, but I'd rather address this by making it better in Stem...

  https://trac.torproject.org/projects/tor/ticket/10107