First stab at reconnection support
This really shouldn't be hard. While writing Stem I took reconnectability into
account from the beginning so it should be as simple as re-connecting and
re-authenticating.

However, something failing badly. Unfortunately this is proving tricky to
troubleshoot so think I'm gonna move on for now, then loop back when more of
arm is rewritten. Hopefully this will make investigating easier.