Rewriting state file parsing
Revising the part of prepopulate_from_state() that parses our state file. Most
interesting bit is that we were doing a dumb method of converting a 'UTC
timestamp => unix time'. We used mktime() which converted to localtime, then
converted back. The calendar.timegm() function, however, does exactly what we
want, as described in the table at the top of...

  https://docs.python.org/3.3/library/time.html