Better handle bandwidth averages
Arm was made in a time before 'GETINFO traffic/*'. In fact, I added those
options just for arm. :)

We had pretty odd handling for both the bandwidth total and average, properly
tracking the former but only basing the later on the time we've been running.

Proper behavior is...

  * Base the average on the whole duration tor's been running if we can get
    both the bandwidth totals and tor's uptime.

  * If not then base both the total and average on the time arm's running.

This is not only more correct and intuitive, but leads to much simpler code
since we don't need to track the prepopulated values separately. Dropping a
couple helper functions while we're at it.