Give daemon tracker test more time
Our daemon tracker checks that when configured to run every 0.01 seconds it
fires at *least* twice in 0.05. As with any timing test though this can still go
awry...

======================================================================
FAIL: test_pausing_daemon (util.tracker.daemon.TestDaemon)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/usr/local/lib/python2.7/dist-packages/mock.py", line 1201, in patched
    return func(*args, **keywargs)
  File "/home/atagar/Desktop/arm/test/util/tracker/daemon.py", line 76, in test_pausing_daemon
    self.assertTrue(2 < daemon.run_counter())
AssertionError: False is not true

----------------------------------------------------------------------

Bumping the test time to 0.2 to give it four times as long. In all the test
runs I've done this is the first time for this to crop up so should be more
than enough.