Moving graph stat inititalization into the graph panel
Hmmmm, not sure why we did this. Our controller initialized graph stats then
added them to the graph panel. Maybe this made sense with a past feature or
pausing implementation.

... or maybe it still makes sense and I'm not spotting why. It's a lot simpler
for the panel to be responsible for its own stats so going with this for now.
Guess we'll see if this bites me down the line.