Moving graph statistical information to a Stat class
Simple class to represent the statistical information on the graph. Yes, we had
a previous GraphStat class but that confused the method that we'd use to fetch
data with both a primary and secondary stat. This doubled the amount of code we
needed. Oh, and the new Stat has a slightly nicer implementation.