Revising ResourceStats class
Nothing much. Main difference is that we only registered stats if we had a new
measurement. This was incorrect since we're supposed to register an update
every second (not be limited by the rate of the tracker).

I dislike the bytes => MB => bytes converting. Both this stat and bandwidth do
this to shorten the y-axis labels but I'm starting to wonder if we should
sacrifice a little data for proper labeling. Something to decide when revising
the GraphPanel itself.